chore(ci): DMND CI/CD actions to sync and deploy upstream tags.
Two GitHub actions. A sync action to check for new upstream tags on a
schedule and then merge the downstream deployment-specific code into a
new downstream tag named `${UPSTREAM_TAG}-dmnd.N` and push it to the
downstream repo.

When a new tag is pushed containing the deploy workflow, the deploy
workflow runs to build an OCI image and deploy it to an AWS ECR.